Put auth service functions to services folder
