+++ app/(auth)/sign-up.tsx
@@ -11,7 +11,8 @@ import AppTitle from '@/components/AppTitle';
 import FQTextInput from '@/components/FQTextInput';
 import FQButton from '@/components/FQButton';
 import { Href, router } from 'expo-router';
-import { isEmailValid, signUp } from '@/utils/auth';
+import { signUp } from '@/services/auth';
+import { isEmailValid } from '@/utils/auth';
 import { SignUpErrorState } from '@/types/auth';
 import { useUserStore } from '@/store/user';
 import { BASE_USER } from '@/constants/user';
